overload operators

+++ opMinus.py
@@ -26,7 +26,7 @@ class OpMinus(OpSpot):
             
         m = A.m
         n = A.n
-        op = OpSpot.__new__(subtype,'Minus',m,n,A-B)
+        op = OpSpot.__new__(subtype,'Minus',m,n,A.double()-B.double())
         op.cflag = A.cflag or B.cflag
         op.linear = A.linear or B.linear
         op.sweepflag = A.sweepflag and B.sweepflag
